+++ rankings/templates/rankings/all_couples.html
@@ -26,7 +26,7 @@
             <td>{{ couple.dancer_1 }}</td>
             <td>and</td>
             <td>{{ couple.dancer_2 }}</td>
-            <td><a href="{% url 'viewcouple' couple.id %}"">{{ couple.get_couple_type_display }}</a></td>
+            <td><a href="{% url 'view_couple' couple.id %}"">{{ couple.get_couple_type_display }}</a></td>
           </tr>
           {% endfor %}
         </tbody>
@@ -35,7 +35,7 @@
       {% include "rankings/pagination.html" %}
 
       <div class="text-center my-3">
-        <a class="btn btn-primary" href="{% url 'createcouple' %}">Add Couple</a>
+        <a class="btn btn-primary" href="{% url 'create_couple' %}">Add Couple</a>
       </div>
     </div>
   </div>
