+++ rankings/views/__init__.py
@@ -1,0 +1,5 @@
+from .calc_rankings import *
+from .couple_crud import *
+from .dancer_crud import *
+from .home import *
+from .user_access import *
